move rd_msg() to new file
